requirements.txt
absl-py==2.2.1
accelerate==1.5.2
ai2-olmo==0.6.0
ai2-olmo-core==0.1.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
backports.tarfile==1.2.0
beaker-gantry==1.14.1
beaker-py==1.34.1
black==23.12.1
boltons==25.0.0
boto3==1.37.25
botocore==1.37.25
build==1.2.2.post1
cached-property==2.0.1
cached_path==1.7.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
chex==0.1.89
click==8.1.8
click-help-colors==0.9.4
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
datasets==3.5.0
decorator==5.2.1
dill==0.3.8
distro==1.9.0
docker==7.1.0
docker-pycreds==0.4.0
docutils==0.21.2
editdistance==0.8.1
einops==0.8.1
einops-exts==0.0.4
etils==1.12.2
exceptiongroup==1.2.2
face==24.0.0
fastapi==0.115.12
ffmpy==0.5.0
fiddle==0.3.0
filelock==3.18.0
flax==0.10.5
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2023.9.2
ftfy==6.3.1
gcsfs==2023.9.2
gitdb==4.0.12
GitPython==3.1.44
glom==24.11.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.176.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.3
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
googletrans==4.0.0rc1
gradio==5.14.0
gradio_client==1.7.0
graphviz==0.20.3
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
huggingface-hub==0.27.1
humanize==4.12.2
hyperframe==5.2.0
id==1.5.0
idna==2.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
iniconfig==2.1.0
isort==5.12.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jax==0.5.3
jaxlib==0.5.3
jeepney==0.9.0
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
keyring==25.6.0
kiwisolver==1.4.8
Levenshtein==0.27.1
libcst==1.7.0
lightning-utilities==0.14.2
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
mdurl==0.1.2
ml_dtypes==0.5.1
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.3.1
multiprocess==0.70.16
mypy==1.3.0
mypy-extensions==1.0.0
narwhals==1.37.1
necessary==0.4.3
nest-asyncio==1.6.0
networkx==3.4.2
nh3==0.2.21
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
openai==1.70.0
opt_einsum==3.4.0
optax==0.2.4
orbax-checkpoint==0.11.10
orjson==3.10.16
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
petname==2.6
pillow==11.1.0
platformdirs==4.3.7
pluggy==1.5.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
pydub==0.25.1
Pygments==2.19.1
pynndescent==0.5.13
pyparsing==3.2.3
pyproject_hooks==1.2.0
pytest==8.3.5
pytest-sphinx==0.6.3
python-dateutil==2.9.0.post0
python-Levenshtein==0.27.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.12.2
readme_renderer==44.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
requirements-parser==0.11.0
rfc3986==1.5.0
rich==13.9.4
rpds-py==0.24.0
rsa==4.9
ruff==0.11.2
s3transfer==0.11.4
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
SecretStorage==3.3.3
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.25.0
setproctitle==1.3.5
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
smart-open==7.1.0
smashed==0.21.5
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.1
sympy==1.13.1
tensorstore==0.1.73
threadpoolctl==3.6.0
tokenizers==0.21.2
tomli==2.2.1
tomlkit==0.13.2
toolz==1.0.0
torch==2.3.1+cu121
torchmetrics==1.7.0
torchvision==0.18.0+cu121
tqdm==4.67.1
transformers==4.48.0
treescope==0.1.9
triton==2.3.1
trouting==0.3.3
twine==6.1.0
typer==0.15.2
types-setuptools==78.1.0.20250329
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
umap-learn==0.5.7
uritemplate==4.2.0
urllib3==2.3.0
uvicorn==0.34.0
wandb==0.19.9
wcwidth==0.2.13
websockets==14.2
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0

# Git dependencies (install separately)
# megablocks @ git+https://github.com/Muennighoff/megablocks.git@4a25bc7b5665bcb9da93d72d5ad0c14d41e1a351
# stanford-stk @ git+https://git@github.com/stanford-futuredata/stk.git@a1ddf98466730b88a2988860a9d8000fd1833301
# ai2_molmo @ git+ssh://git@github.com/BennoKrojer/molmo.git@77116b18685169d532fc7dcf61ffc03cc22fdc03#egg=ai2_molmo&subdirectory=../../../third_party/molmo
